+++ src/onnx_nn_in_c/c_wrapper.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <constants.h>
 
 void load_model(const char* model_path, int core_count);
-double* run_zernike_model(double input_pixels[IPS][IPS]);
+float* run_zernike_model(float input_pixels[IPS][IPS]);
 void close_model();
 
 #ifdef __cplusplus
